Fix a bug into fobos local variables parsing
Fix a bug into fobos local variables parsing regular expression.

Why:

There was a bug on the regular expression used for parsing the *local
varaibles* defined into fobos file. With the bugged FoBiS version it
happens that for a fobos similar to

```ini
[common]
$INT_DEB = blablabla
$INT_DEB_PURE = foofoofoo

[tests]
cflags = $INT_DEB_PURE
```

the `cflags` is wrongly expanded to `cflags = blablabla_PURE`, namely
the `$INT_DEB_PURE` is wrongly parsed. With the new version this bug
seems to be fixed and the result is `cflags = foofoofoo`.

This change addresses the need by:

Correct the regular expression of the Fobos class used for parsing local
variables.

Side effects:

Nothing.