Implement support for submodules
Implement support for submodules.

Submodules are *strange*: a parent module of a submodule do not have an
explicit dependency on the submodule. Thus submodules must be treated as
old-fashioned non-module libraries: all submodules found must be
compiled and passed to linker as it happens for non-module libraries.
Submodules are explicitely dependent on their parent modules that must
be compiled before submodules themself.