Quick fix for doctests inside include file sources
Quick fix for doctests inside include file sources: using the
`#include` preprocess macro to achieve templating breaks doctests...
quick and dirty fix has be to preprocess (by cpp) sources before parsing
for doctests. It is necessary a better fix, but for now it works (if cpp
is present in PATH).